Refactor vision dynamics to new module
